update scripts to incorporate new url format.
URLs are now unique/specific to each date. In addition, the date contained within the url is actually a day ahead of the as of date for the data contained within. For example, a url of 04292020 will contain data as of 04282020. Changed the code to reflect this update, and to ensure that dates are accurately recorded in the dataframe itself.